+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - utils : make `writeTextureToVideoFrame` a member function (https://github.com/Simple-Robotics/candlewick/pull/62)
 - utils : rename `writeToFile` to `saveTextureToFile` (https://github.com/Simple-Robotics/candlewick/pull/62)
 - utils/PixelFormatConversion.h : make bgra conversion function in-place (https://github.com/Simple-Robotics/candlewick/pull/63)
+- utils & third-party : switch from `stb_image_write.h` to `fpng` for writing PNG files (https://github.com/Simple-Robotics/candlewick/pull/64)
 
 ### Fixed